feat(dashboard): split UnauthenticatedLayout into AuthFlowLayout and GuestGuard
AuthFlowLayout is the dark two-column sign-in shell and nothing else.
GuestGuard redirects signed-in visitors to their post-sign-in target.
/password/reset renders without the guard instead of the layout checking
the route, since it is reached while signed in via the reset token.
