+++ dashboard/src/components/layout/AuthFlowLayout/index.ts
@@ -1,0 +1,2 @@
+export type { AuthFlowLayoutProps } from '@/components/layout/AuthFlowLayout/AuthFlowLayout';
+export { default as AuthFlowLayout } from '@/components/layout/AuthFlowLayout/AuthFlowLayout';
