--- dashboard/src/components/layout/UnauthenticatedLayout/index.ts
@@ -1,2 +1,0 @@
-export * from './UnauthenticatedLayout';
-export { default as UnauthenticatedLayout } from './UnauthenticatedLayout';
