+++ dashboard/src/features/auth/GuestGuard/index.ts
@@ -1,0 +1,1 @@
+export { default as GuestGuard } from '@/features/auth/GuestGuard/GuestGuard';
