+++ dashboard/src/pages/password/new.tsx
@@ -7,8 +7,9 @@ import * as Yup from 'yup';
 import { NavLink } from '@/components/common/NavLink';
 import { Form } from '@/components/form/Form';
 import { FormInput } from '@/components/form/FormInput';
-import { UnauthenticatedLayout } from '@/components/layout/UnauthenticatedLayout';
+import { AuthFlowLayout } from '@/components/layout/AuthFlowLayout';
 import { ButtonWithLoading } from '@/components/ui/v3/button';
+import { GuestGuard } from '@/features/auth/GuestGuard';
 import { appendPkceId, generateAndStorePKCE } from '@/lib/pkce';
 import { useNhostClient } from '@/providers/nhost';
 import { getToastStyleProps } from '@/utils/constants/settings';
@@ -157,8 +158,8 @@ export default function NewPasswordPage() {
 
 NewPasswordPage.getLayout = function getLayout(page: ReactElement) {
   return (
-    <UnauthenticatedLayout title="Request Password Reset">
-      {page}
-    </UnauthenticatedLayout>
+    <AuthFlowLayout title="Request Password Reset">
+      <GuestGuard>{page}</GuestGuard>
+    </AuthFlowLayout>
   );
 };
