+++ dashboard/src/pages/password/reset.tsx
@@ -6,7 +6,7 @@ import * as Yup from 'yup';
 import { NavLink } from '@/components/common/NavLink';
 import { Form } from '@/components/form/Form';
 import { FormInput } from '@/components/form/FormInput';
-import { UnauthenticatedLayout } from '@/components/layout/UnauthenticatedLayout';
+import { AuthFlowLayout } from '@/components/layout/AuthFlowLayout';
 import { ButtonWithLoading } from '@/components/ui/v3/button';
 import useActionWithElevatedPermissions from '@/features/account/settings/hooks/useActionWithElevatedPermissions';
 import { useNhostClient } from '@/providers/nhost';
@@ -106,8 +106,8 @@ export default function ResetPasswordPage() {
 
 ResetPasswordPage.getLayout = function getLayout(page: ReactElement) {
   return (
-    <UnauthenticatedLayout title="Request Password Reset">
+    <AuthFlowLayout title="Request Password Reset">
       {page}
-    </UnauthenticatedLayout>
+    </AuthFlowLayout>
   );
 };
