+++ dashboard/src/pages/signin.tsx
@@ -1,9 +1,10 @@
 import NextLink from 'next/link';
 import { type ReactElement, useEffect } from 'react';
 import { SignInRightColumn } from '@/components/auth/SignInRightColumn';
-import { UnauthenticatedLayout } from '@/components/layout/UnauthenticatedLayout';
+import { AuthFlowLayout } from '@/components/layout/AuthFlowLayout';
 import { Button } from '@/components/ui/v3/button';
 import { Separator } from '@/components/ui/v3/separator';
+import { GuestGuard } from '@/features/auth/GuestGuard';
 import { SignInWithSecurityKey } from '@/features/auth/SignIn/SecurityKey';
 import { SignInWithGithub } from '@/features/auth/SignIn/SignInWithGithub';
 import { useAuth } from '@/providers/Auth';
@@ -80,11 +81,11 @@ export default function SigninPage() {
 
 SigninPage.getLayout = function getLayout(page: ReactElement) {
   return (
-    <UnauthenticatedLayout
+    <AuthFlowLayout
       title="Sign In"
       rightColumnContent={<SignInRightColumn />}
     >
-      {page}
-    </UnauthenticatedLayout>
+      <GuestGuard>{page}</GuestGuard>
+    </AuthFlowLayout>
   );
 };
