+++ dashboard/src/pages/signup.tsx
@@ -3,8 +3,9 @@ import NextLink from 'next/link';
 import type { ReactElement } from 'react';
 import { useCallback } from 'react';
 import { CookieConsent } from '@/components/common/CookieConsent';
-import { UnauthenticatedLayout } from '@/components/layout/UnauthenticatedLayout';
+import { AuthFlowLayout } from '@/components/layout/AuthFlowLayout';
 import { Separator } from '@/components/ui/v3/separator';
+import { GuestGuard } from '@/features/auth/GuestGuard';
 import { SignUpTabs } from '@/features/auth/SignUp/SignUpTabs';
 import { SignUpWithGithub } from '@/features/auth/SignUp/SignUpWithGithub';
 
@@ -183,11 +184,11 @@ export default function SignUpPage() {
 
 SignUpPage.getLayout = function getLayout(page: ReactElement) {
   return (
-    <UnauthenticatedLayout
+    <AuthFlowLayout
       title="Sign Up"
       rightColumnContent={rightColumnContent}
     >
-      {page}
-    </UnauthenticatedLayout>
+      <GuestGuard>{page}</GuestGuard>
+    </AuthFlowLayout>
   );
 };
